chore: Use new `getArtist` query for "Current Artist" screen
