+++ mobile/src/data/artist/types.ts
@@ -5,3 +5,10 @@ export type ArtistTrack = {
   duration: number;
   album: string | null;
 };
+
+export type ArtistAlbum = {
+  id: string;
+  name: string;
+  artwork: string | null;
+  year: string;
+};
