+++ mobile/src/queries/keyStore.ts
@@ -12,7 +12,6 @@ import {
 } from "~/db/schema";
 
 import { getAlbum, getAlbums } from "~/api/album";
-import { getArtistAlbums } from "~/api/artist";
 import { getFolder } from "~/api/folder";
 import { getPlaylist, getPlaylists } from "~/api/playlist";
 import {
@@ -25,6 +24,7 @@ import {
   getTrackGenres,
   getTrackPlaylists,
 } from "~/api/track";
+import { getArtist } from "~/data/artist/api";
 import { getGenre, getGenresSummary } from "~/data/genre/api";
 
 import { iAsc, throwIfNoResults } from "~/lib/drizzle";
@@ -81,31 +81,7 @@ export const queries = createQueryKeyStore({
     },
     detail: (artistName: string) => ({
       queryKey: [artistName],
-      queryFn: async () => {
-        const [artistData, artistTracks, artistAlbums] = await Promise.all([
-          throwIfNoResults(
-            db.query.artists.findFirst({
-              where: (fields, { eq }) => eq(fields.name, artistName),
-            }),
-          ),
-          //? Get the tracks associated with the artist in alphabetical order.
-          db
-            .select({
-              id: tracks.id,
-              name: tracks.name,
-              artwork: tracks.artwork,
-              duration: tracks.duration,
-              album: { name: albums.name, artwork: albums.artwork },
-            })
-            .from(tracksToArtists)
-            .where(eq(tracksToArtists.artistName, artistName))
-            .innerJoin(tracks, eq(tracksToArtists.trackId, tracks.id))
-            .leftJoin(albums, eq(tracks.albumId, albums.id))
-            .orderBy(iAsc(tracks.name)),
-          getArtistAlbums(artistName),
-        ]);
-        return { ...artistData, tracks: artistTracks, albums: artistAlbums };
-      },
+      queryFn: () => getArtist(artistName),
     }),
   },
   /** Query keys used in `useQuery` for favorite media. */
